requirements.txt
# Python 3.12.7
openai==2.15.0
click==8.3.1
rich==14.3.1
tiktoken==0.12.0
pydantic==2.12.5
black==26.1.0
tomli==2.4.0
python-dotenv==1.2.1